refactor/home-view: Improvement in the code and added documentation for this UI
