+++ mobile_dev/lib/main.dart
@@ -11,7 +11,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: const BoardingView(),
+      home: const OnBoardingView(),
     );
   }
 }
