--- mobile_dev/lib/src/views/boarding/boarding_view.dart
@@ -1,53 +1,0 @@
-import 'package:flutter/material.dart';
-
-class BoardingView extends StatelessWidget {
-  const BoardingView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Colors.blue,
-      child: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            const SizedBox(height: 30),
-            const Image(
-              image: AssetImage('assets/health-insurance.png'),
-              height: 300,
-            ),
-            const SizedBox(height: 20),
-            const Text(
-              '¡Bienvenido a Hospinova!',
-              style: TextStyle(
-                color: Colors.white,
-                fontSize: 24,
-                fontWeight: FontWeight.bold,
-                decoration: TextDecoration.none,
-              ),
-            ),
-            const SizedBox(height: 20),
-            ElevatedButton(
-              onPressed: () {},
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.white,
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 50,
-                  vertical: 15,
-                ),
-              ),
-              child: const Text(
-                'EMPEZAR',
-                style: TextStyle(
-                  color: Colors.blue,
-                  fontSize: 18,
-                ),
-              ),
-            ),
-            const SizedBox(height: 50),
-          ],
-        ),
-      ),
-    );
-  }
-}
