+++ mobile_dev/lib/src/views/views.dart
@@ -1,1 +1,1 @@
-export 'boarding/boarding_view.dart';
+export 'onboarding/on_boarding_view.dart';
